More accurate model scaling
First tried getting bounding box in screen coordinates but that went crazy when the model went into the screen. Using clip space coordinates mostly works but the viewport scaling was fighting the depth scaling on some models causing super jitter. Somehow it was fixed by adjusting the origin depth to the middle of the frustrum and setting a lower limit (0.5f) on the max depth coord. I don't understand it really.